Refactor input and camera types with improved enum string conversions and float16 usage
- Added INI_START_COMMENT_PREFIXES to build.sh INI defines.
- Switch camera_t zoom fields to float16_t for improved precision and memory usage.
- Removed unused camera_t load/unload functions; updated camera_t$update signature.
- Enabled building tests for input and camera_t in config.sh.
- Fixed includes in controls_t to use "input_t.h" instead of "input.h".
- Revamped button_t and direction_t enums:
  - Added char constants and macros for button and direction characters.
  - Replaced previous conversion functions with static inline functions returning static strings.
  - Added support for multiple simultaneous buttons/directions by concatenating chars.
- Removed deprecated input.h, now only input_t.h is used.
- Changed inputBuffer_t currentBufferIndex to size_t.
- Updated input_t$convert$toString to a static inline returning a static string combining direction and buttons.

Refactor input handling and update player API

- Remove obsolete input.c and input_t.c files, cleaning up unused input code.
- Update inputBuffer_t.c:
  * Refine input buffer index types and buffer length handling.
  * Implement partial input sequence retrieval with frame timing logic.
  * Add TODO comments for further frame and input sequence improvements.
- Modify event.c to replace input.h include with inputBuffer_t.h.
- Remove deprecated player_t$move() function from player_t interface and implementation.
- Enhance stdfunc.h with additional native array utility macros and reverse iteration macros.
- Add basic test setup and tests for camera_t module.

Add comprehensive unit tests for input and inputBuffer modules

- Introduce tests for direction_t and button_t conversion to static strings
- Add tests for input_t macros and conversion functions (GET_DIRECTION, GET_BUTTON, MAKE_INPUT, toStaticString)
- Implement inputBuffer_t tests: create, destroy, insert, and inputsSequence$get behavior
- Cover edge cases including invalid inputs and buffer wrapping scenarios
- Include TODO placeholders for unfinished test implementations
- Add shell script to define FILES_TO_COMPILE environment variable for tests

Add window_t destroy tests and refactor vsync type definitions

- Add unit tests for window_t$destroy, covering NULL and valid window cases
- Move vsync_t enum and related macros from vsync.h to new vsync_t.h header
- Update vsync.h to include vsync_t.h, cleaning up enum and macro declarations

+++ input/include/inputBuffer_t.h
@@ -15,7 +15,7 @@
 typedef struct {
     input_t inputs[ INPUT_BUFFER_LENGTH ];
     size_t frames[ INPUT_BUFFER_LENGTH ];
-    uint8_t currentBufferIndex;
+    size_t currentBufferIndex;
 } inputBuffer_t;
 
 inputBuffer_t inputBuffer_t$create( void );
